[TASK] Update Ext JS to version 4
Included Ext JS 4. Modified the viewhelper so it works with version 4.
Copied the old viewhelper to IncludeVersion3ViewHelper. It now needs a
path to a standard Ext JS 3 directory, because version 3 will not be
shipped with this package anymore.
The Extdirect API stays the same as apparently nothing changed from
Ext 3 to Ext 4.

Change-Id: Icd1bf7e79eaee02c2702d8c94257f58c2232d2f9